requirements.txt
fastmcp
google-auth
python-dotenv
requests
google-cloud-storage
PyPDF2
